trying a different way to run the data_file_directory_uploader
